[Gold III] Title: 웜홀, Time: 28 ms, Memory: 2168 KB -BaekjoonHub

+++ 백준/Gold/1865. 웜홀/웜홀.cc
@@ -5,7 +5,7 @@
 
 using namespace std;
 typedef long long ll;
-ll dist[501];
+int dist[501];
 
 vector<pair<int, int>> v[501];
